Modify the active states behavior
Now the when the share button is pressed the social media links will be
using position: absolute to be on top of the author-info div. This
instead of using the display: none property to hide the author-info div
and show the social-media-container.

The necessary changes to the HTML and CSS for this to behave correctly
were made.